First round to Address Inconsistencies in postgresql Database schema

+++ installation/sql/postgresql/joomla.sql
@@ -88,7 +88,7 @@ SELECT setval('#__assets_id_seq', 55, false);
 -- Table: #__associations
 --
 CREATE TABLE "#__associations" (
-  "id" int NOT NULL,
+  "id" bigint NOT NULL,
   "context" varchar(50) NOT NULL,
   "key" char(32) NOT NULL,
   CONSTRAINT "#__associations_idx_context_id" PRIMARY KEY ("context", "id")
@@ -415,11 +415,11 @@ SELECT setval('#__content_types_type_id_seq', 10000, false);
 --
 CREATE TABLE "#__contentitem_tag_map" (
   "type_alias" varchar(255) NOT NULL DEFAULT '',
-  "core_content_id" integer NOT NULL,
-  "content_item_id" integer NOT NULL,
-  "tag_id" integer NOT NULL,
+  "core_content_id" bigint NOT NULL,
+  "content_item_id" bigint NOT NULL,
+  "tag_id" bigint NOT NULL,
   "tag_date" timestamp without time zone DEFAULT '1970-01-01 00:00:00' NOT NULL,
-  "type_id" integer NOT NULL,
+  "type_id" bigint NOT NULL,
  CONSTRAINT "#__uc_ItemnameTagid" UNIQUE ("type_id", "content_item_id", "tag_id")
 );
 CREATE INDEX "#__contentitem_tag_map_idx_tag_type" ON "#__contentitem_tag_map" ("tag_id", "type_id");
@@ -646,13 +646,13 @@ CREATE TABLE "#__finder_filters" (
   "alias" varchar(255) NOT NULL,
   "state" smallint DEFAULT 1 NOT NULL,
   "created" timestamp without time zone DEFAULT '1970-01-01 00:00:00' NOT NULL,
-  "created_by" integer NOT NULL,
+  "created_by" bigint NOT NULL,
   "created_by_alias" varchar(255) NOT NULL,
   "modified" timestamp without time zone DEFAULT '1970-01-01 00:00:00' NOT NULL,
-  "modified_by" integer DEFAULT 0 NOT NULL,
-  "checked_out" integer DEFAULT 0 NOT NULL,
+  "modified_by" bigint DEFAULT 0 NOT NULL,
+  "checked_out" bigint DEFAULT 0 NOT NULL,
   "checked_out_time" timestamp without time zone DEFAULT '1970-01-01 00:00:00' NOT NULL,
-  "map_count" integer DEFAULT 0 NOT NULL,
+  "map_count" bigint DEFAULT 0 NOT NULL,
   "data" text NOT NULL,
   "params" text,
   PRIMARY KEY ("filter_id")
@@ -671,7 +671,7 @@ CREATE TABLE "#__finder_links" (
   "md5sum" varchar(32) DEFAULT NULL,
   "published" smallint DEFAULT 1 NOT NULL,
   "state" integer DEFAULT 1,
-  "access" integer DEFAULT 0,
+  "access" bigint DEFAULT 0,
   "language" varchar(8) DEFAULT '' NOT NULL,
   "publish_start_date" timestamp without time zone DEFAULT '1970-01-01 00:00:00' NOT NULL,
   "publish_end_date" timestamp without time zone DEFAULT '1970-01-01 00:00:00' NOT NULL,
@@ -694,8 +694,8 @@ CREATE INDEX "#__finder_links_idx_published_sale" on "#__finder_links" ("publish
 -- Table: #__finder_links_terms0
 --
 CREATE TABLE "#__finder_links_terms0" (
-  "link_id" integer NOT NULL,
-  "term_id" integer NOT NULL,
+  "link_id" bigint NOT NULL,
+  "term_id" bigint NOT NULL,
   "weight" numeric(8,2) NOT NULL,
   PRIMARY KEY ("link_id", "term_id")
 );
@@ -706,8 +706,8 @@ CREATE INDEX "#__finder_links_terms0_idx_link_term_weight" on "#__finder_links_t
 -- Table: #__finder_links_terms1
 --
 CREATE TABLE "#__finder_links_terms1" (
-  "link_id" integer NOT NULL,
-  "term_id" integer NOT NULL,
+  "link_id" bigint NOT NULL,
+  "term_id" bigint NOT NULL,
   "weight" numeric(8,2) NOT NULL,
   PRIMARY KEY ("link_id", "term_id")
 );
@@ -718,8 +718,8 @@ CREATE INDEX "#__finder_links_terms1_idx_link_term_weight" on "#__finder_links_t
 -- Table: #__finder_links_terms2
 --
 CREATE TABLE "#__finder_links_terms2" (
-  "link_id" integer NOT NULL,
-  "term_id" integer NOT NULL,
+  "link_id" bigint NOT NULL,
+  "term_id" bigint NOT NULL,
   "weight" numeric(8,2) NOT NULL,
   PRIMARY KEY ("link_id", "term_id")
 );
@@ -730,8 +730,8 @@ CREATE INDEX "#__finder_links_terms2_idx_link_term_weight" on "#__finder_links_t
 -- Table: #__finder_links_terms3
 --
 CREATE TABLE "#__finder_links_terms3" (
-  "link_id" integer NOT NULL,
-  "term_id" integer NOT NULL,
+  "link_id" bigint NOT NULL,
+  "term_id" bigint NOT NULL,
   "weight" numeric(8,2) NOT NULL,
   PRIMARY KEY ("link_id", "term_id")
 );
@@ -742,8 +742,8 @@ CREATE INDEX "#__finder_links_terms3_idx_link_term_weight" on "#__finder_links_t
 -- Table: #__finder_links_terms4
 --
 CREATE TABLE "#__finder_links_terms4" (
-  "link_id" integer NOT NULL,
-  "term_id" integer NOT NULL,
+  "link_id" bigint NOT NULL,
+  "term_id" bigint NOT NULL,
   "weight" numeric(8,2) NOT NULL,
   PRIMARY KEY ("link_id", "term_id")
 );
@@ -754,8 +754,8 @@ CREATE INDEX "#__finder_links_terms4_idx_link_term_weight" on "#__finder_links_t
 -- Table: #__finder_links_terms5
 --
 CREATE TABLE "#__finder_links_terms5" (
-  "link_id" integer NOT NULL,
-  "term_id" integer NOT NULL,
+  "link_id" bigint NOT NULL,
+  "term_id" bigint NOT NULL,
   "weight" numeric(8,2) NOT NULL,
   PRIMARY KEY ("link_id", "term_id")
 );
@@ -766,8 +766,8 @@ CREATE INDEX "#__finder_links_terms5_idx_link_term_weight" on "#__finder_links_t
 -- Table: #__finder_links_terms6
 --
 CREATE TABLE "#__finder_links_terms6" (
-  "link_id" integer NOT NULL,
-  "term_id" integer NOT NULL,
+  "link_id" bigint NOT NULL,
+  "term_id" bigint NOT NULL,
   "weight" numeric(8,2) NOT NULL,
   PRIMARY KEY ("link_id", "term_id")
 );
@@ -778,8 +778,8 @@ CREATE INDEX "#__finder_links_terms6_idx_link_term_weight" on "#__finder_links_t
 -- Table: #__finder_links_terms7
 --
 CREATE TABLE "#__finder_links_terms7" (
-  "link_id" integer NOT NULL,
-  "term_id" integer NOT NULL,
+  "link_id" bigint NOT NULL,
+  "term_id" bigint NOT NULL,
   "weight" numeric(8,2) NOT NULL,
   PRIMARY KEY ("link_id", "term_id")
 );
@@ -790,8 +790,8 @@ CREATE INDEX "#__finder_links_terms7_idx_link_term_weight" on "#__finder_links_t
 -- Table: #__finder_links_terms8
 --
 CREATE TABLE "#__finder_links_terms8" (
-  "link_id" integer NOT NULL,
-  "term_id" integer NOT NULL,
+  "link_id" bigint NOT NULL,
+  "term_id" bigint NOT NULL,
   "weight" numeric(8,2) NOT NULL,
   PRIMARY KEY ("link_id", "term_id")
 );
@@ -802,8 +802,8 @@ CREATE INDEX "#__finder_links_terms8_idx_link_term_weight" on "#__finder_links_t
 -- Table: #__finder_links_terms9
 --
 CREATE TABLE "#__finder_links_terms9" (
-  "link_id" integer NOT NULL,
-  "term_id" integer NOT NULL,
+  "link_id" bigint NOT NULL,
+  "term_id" bigint NOT NULL,
   "weight" numeric(8,2) NOT NULL,
   PRIMARY KEY ("link_id", "term_id")
 );
@@ -814,8 +814,8 @@ CREATE INDEX "#__finder_links_terms9_idx_link_term_weight" on "#__finder_links_t
 -- Table: #__finder_links_termsa
 --
 CREATE TABLE "#__finder_links_termsa" (
-  "link_id" integer NOT NULL,
-  "term_id" integer NOT NULL,
+  "link_id" bigint NOT NULL,
+  "term_id" bigint NOT NULL,
   "weight" numeric(8,2) NOT NULL,
   PRIMARY KEY ("link_id", "term_id")
 );
@@ -826,8 +826,8 @@ CREATE INDEX "#__finder_links_termsa_idx_link_term_weight" on "#__finder_links_t
 -- Table: #__finder_links_termsb
 --
 CREATE TABLE "#__finder_links_termsb" (
-  "link_id" integer NOT NULL,
-  "term_id" integer NOT NULL,
+  "link_id" bigint NOT NULL,
+  "term_id" bigint NOT NULL,
   "weight" numeric(8,2) NOT NULL,
   PRIMARY KEY ("link_id", "term_id")
 );
@@ -838,8 +838,8 @@ CREATE INDEX "#__finder_links_termsb_idx_link_term_weight" on "#__finder_links_t
 -- Table: #__finder_links_termsc
 --
 CREATE TABLE "#__finder_links_termsc" (
-  "link_id" integer NOT NULL,
-  "term_id" integer NOT NULL,
+  "link_id" bigint NOT NULL,
+  "term_id" bigint NOT NULL,
   "weight" numeric(8,2) NOT NULL,
   PRIMARY KEY ("link_id", "term_id")
 );
@@ -850,8 +850,8 @@ CREATE INDEX "#__finder_links_termsc_idx_link_term_weight" on "#__finder_links_t
 -- Table: #__finder_links_termsd
 --
 CREATE TABLE "#__finder_links_termsd" (
-  "link_id" integer NOT NULL,
-  "term_id" integer NOT NULL,
+  "link_id" bigint NOT NULL,
+  "term_id" bigint NOT NULL,
   "weight" numeric(8,2) NOT NULL,
   PRIMARY KEY ("link_id", "term_id")
 );
@@ -862,8 +862,8 @@ CREATE INDEX "#__finder_links_termsd_idx_link_term_weight" on "#__finder_links_t
 -- Table: #__finder_links_termse
 --
 CREATE TABLE "#__finder_links_termse" (
-  "link_id" integer NOT NULL,
-  "term_id" integer NOT NULL,
+  "link_id" bigint NOT NULL,
+  "term_id" bigint NOT NULL,
   "weight" numeric(8,2) NOT NULL,
   PRIMARY KEY ("link_id", "term_id")
 );
@@ -874,8 +874,8 @@ CREATE INDEX "#__finder_links_termse_idx_link_term_weight" on "#__finder_links_t
 -- Table: #__finder_links_termsf
 --
 CREATE TABLE "#__finder_links_termsf" (
-  "link_id" integer NOT NULL,
-  "term_id" integer NOT NULL,
+  "link_id" bigint NOT NULL,
+  "term_id" bigint NOT NULL,
   "weight" numeric(8,2) NOT NULL,
   PRIMARY KEY ("link_id", "term_id")
 );
@@ -887,10 +887,10 @@ CREATE INDEX "#__finder_links_termsf_idx_link_term_weight" on "#__finder_links_t
 --
 CREATE TABLE "#__finder_taxonomy" (
   "id" serial NOT NULL,
-  "parent_id" integer DEFAULT 0 NOT NULL,
+  "parent_id" bigint DEFAULT 0 NOT NULL,
   "title" varchar(255) NOT NULL,
   "state" smallint DEFAULT 1 NOT NULL,
-  "access" smallint DEFAULT 0 NOT NULL,
+  "access" bigint DEFAULT 0 NOT NULL,
   "ordering" smallint DEFAULT 0 NOT NULL,
   PRIMARY KEY ("id")
 );
@@ -912,8 +912,8 @@ SELECT setval('#__finder_taxonomy_id_seq', 2, false);
 -- Table: #__finder_taxonomy_map
 --
 CREATE TABLE "#__finder_taxonomy_map" (
-  "link_id" integer NOT NULL,
-  "node_id" integer NOT NULL,
+  "link_id" bigint NOT NULL,
+  "node_id" bigint NOT NULL,
   PRIMARY KEY ("link_id", "node_id")
 );
 CREATE INDEX "#__finder_taxonomy_map_link_id" on "#__finder_taxonomy_map" ("link_id");
@@ -1089,7 +1089,7 @@ CREATE INDEX "#__finder_tokens_idx_context" on "#__finder_tokens" ("context");
 -- Table: #__finder_tokens_aggregate
 --
 CREATE TABLE "#__finder_tokens_aggregate" (
-  "term_id" integer NOT NULL,
+  "term_id" bigint NOT NULL,
   "map_suffix" varchar(1) NOT NULL,
   "term" varchar(75) NOT NULL,
   "stem" varchar(75) NOT NULL,
@@ -1131,7 +1131,7 @@ CREATE TABLE "#__languages" (
   "metadesc" text NOT NULL,
   "sitename" varchar(1024) DEFAULT '' NOT NULL,
   "published" bigint DEFAULT 0 NOT NULL,
-  "access" integer DEFAULT 0 NOT NULL,
+  "access" bigint DEFAULT 0 NOT NULL,
   "ordering" bigint DEFAULT 0 NOT NULL,
   PRIMARY KEY ("lang_id"),
   CONSTRAINT "#__languages_idx_sef" UNIQUE ("sef"),
@@ -1162,15 +1162,15 @@ CREATE TABLE "#__menu" (
   "link" varchar(1024) NOT NULL,
   "type" varchar(16) NOT NULL,
   "published" smallint DEFAULT 0 NOT NULL,
-  "parent_id" integer DEFAULT 1 NOT NULL,
-  "level" integer DEFAULT 0 NOT NULL,
-  "component_id" integer DEFAULT 0 NOT NULL,
-  "checked_out" integer DEFAULT 0 NOT NULL,
+  "parent_id" bigint DEFAULT 1 NOT NULL,
+  "level" bigint DEFAULT 0 NOT NULL,
+  "component_id" bigint DEFAULT 0 NOT NULL,
+  "checked_out" bigint DEFAULT 0 NOT NULL,
   "checked_out_time" timestamp without time zone DEFAULT '1970-01-01 00:00:00' NOT NULL,
   "browserNav" smallint DEFAULT 0 NOT NULL,
   "access" bigint DEFAULT 0 NOT NULL,
   "img" varchar(255) DEFAULT '' NOT NULL,
-  "template_style_id" integer DEFAULT 0 NOT NULL,
+  "template_style_id" bigint DEFAULT 0 NOT NULL,
   "params" text DEFAULT '' NOT NULL,
   "lft" bigint DEFAULT 0 NOT NULL,
   "rgt" bigint DEFAULT 0 NOT NULL,
@@ -1294,7 +1294,7 @@ CREATE TABLE "#__modules" (
   "content" text DEFAULT '' NOT NULL,
   "ordering" bigint DEFAULT 0 NOT NULL,
   "position" varchar(50) DEFAULT '' NOT NULL,
-  "checked_out" integer DEFAULT 0 NOT NULL,
+  "checked_out" bigint DEFAULT 0 NOT NULL,
   "checked_out_time" timestamp without time zone DEFAULT '1970-01-01 00:00:00' NOT NULL,
   "publish_up" timestamp without time zone DEFAULT '1970-01-01 00:00:00' NOT NULL,
   "publish_down" timestamp without time zone DEFAULT '1970-01-01 00:00:00' NOT NULL,
@@ -1376,7 +1376,7 @@ CREATE TABLE "#__newsfeeds" (
   "published" smallint DEFAULT 0 NOT NULL,
   "numarticles" bigint DEFAULT 1 NOT NULL,
   "cache_time" bigint DEFAULT 3600 NOT NULL,
-  "checked_out" integer DEFAULT 0 NOT NULL,
+  "checked_out" bigint DEFAULT 0 NOT NULL,
   "checked_out_time" timestamp without time zone DEFAULT '1970-01-01 00:00:00' NOT NULL,
   "ordering" bigint DEFAULT 0 NOT NULL,
   "rtl" smallint DEFAULT 0 NOT NULL,
@@ -1384,10 +1384,10 @@ CREATE TABLE "#__newsfeeds" (
   "language" varchar(7) DEFAULT '' NOT NULL,
   "params" text NOT NULL,
   "created" timestamp without time zone DEFAULT '1970-01-01 00:00:00' NOT NULL,
-  "created_by" integer DEFAULT 0 NOT NULL,
+  "created_by" bigint DEFAULT 0 NOT NULL,
   "created_by_alias" varchar(255) DEFAULT '' NOT NULL,
   "modified" timestamp without time zone DEFAULT '1970-01-01 00:00:00' NOT NULL,
-  "modified_by" integer DEFAULT 0 NOT NULL,
+  "modified_by" bigint DEFAULT 0 NOT NULL,
   "metakey" text NOT NULL,
   "metadesc" text NOT NULL,
   "metadata" text NOT NULL,
@@ -1527,14 +1527,14 @@ CREATE TABLE "#__tags" (
   "metadesc" varchar(1024) NOT NULL,
   "metakey" varchar(1024) NOT NULL,
   "metadata" varchar(2048) NOT NULL,
-  "created_user_id" integer DEFAULT 0 NOT NULL,
+  "created_user_id" bigint DEFAULT 0 NOT NULL,
   "created_time" timestamp without time zone DEFAULT '1970-01-01 00:00:00' NOT NULL,
   "created_by_alias" varchar(255) DEFAULT '' NOT NULL,
-  "modified_user_id" integer DEFAULT 0 NOT NULL,
+  "modified_user_id" bigint DEFAULT 0 NOT NULL,
   "modified_time" timestamp without time zone DEFAULT '1970-01-01 00:00:00' NOT NULL,
   "images" text NOT NULL,
   "urls" text NOT NULL,
-  "hits" integer DEFAULT 0 NOT NULL,
+  "hits" bigint DEFAULT 0 NOT NULL,
   "language" varchar(7) DEFAULT '' NOT NULL,
   "version" bigint DEFAULT 1 NOT NULL,
   "publish_up" timestamp without time zone DEFAULT '1970-01-01 00:00:00' NOT NULL,
@@ -1655,11 +1655,11 @@ CREATE INDEX "#__ucm_content_idx_core_type_id" ON "#__ucm_content" ("core_type_i
 --
 CREATE TABLE "#__ucm_history" (
   "version_id" serial NOT NULL,
-  "ucm_item_id" integer NOT NULL,
-  "ucm_type_id" integer NOT NULL,
+  "ucm_item_id" bigint NOT NULL,
+  "ucm_type_id" bigint NOT NULL,
   "version_note" varchar(255) NOT NULL DEFAULT '',
   "save_date" timestamp with time zone NOT NULL DEFAULT '1970-01-01 00:00:00',
-  "editor_user_id" integer  NOT NULL DEFAULT 0,
+  "editor_user_id" bigint  NOT NULL DEFAULT 0,
   "character_count" integer  NOT NULL DEFAULT 0,
   "sha1_hash" varchar(50) NOT NULL DEFAULT '',
   "version_data" text NOT NULL,
@@ -1848,16 +1848,16 @@ CREATE INDEX "#__user_keys_idx_user_id" ON "#__user_keys" ("user_id");
 --
 CREATE TABLE "#__user_notes" (
   "id" serial NOT NULL,
-  "user_id" integer DEFAULT 0 NOT NULL,
-  "catid" integer DEFAULT 0 NOT NULL,
+  "user_id" bigint DEFAULT 0 NOT NULL,
+  "catid" bigint DEFAULT 0 NOT NULL,
   "subject" varchar(100) DEFAULT '' NOT NULL,
   "body" text NOT NULL,
   "state" smallint DEFAULT 0 NOT NULL,
-  "checked_out" integer DEFAULT 0 NOT NULL,
+  "checked_out" bigint DEFAULT 0 NOT NULL,
   "checked_out_time" timestamp without time zone DEFAULT '1970-01-01 00:00:00' NOT NULL,
-  "created_user_id" integer DEFAULT 0 NOT NULL,
+  "created_user_id" bigint DEFAULT 0 NOT NULL,
   "created_time" timestamp without time zone DEFAULT '1970-01-01 00:00:00' NOT NULL,
-  "modified_user_id" integer NOT NULL,
+  "modified_user_id" bigint NOT NULL,
   "modified_time" timestamp without time zone DEFAULT '1970-01-01 00:00:00' NOT NULL,
   "review_time" timestamp without time zone DEFAULT '1970-01-01 00:00:00' NOT NULL,
   "publish_up" timestamp without time zone DEFAULT '1970-01-01 00:00:00' NOT NULL,
